separate communities data from code
